feat: move clickhandler functionality into Dropdown
